requirements.txt
aiogram==2.25.2
python-dotenv
pillow
google-cloud-firestore
aiohttp
pytz
flask
google-api-python-client
google-auth
